refactor: 404 page migration to Next.js refactor: ErrorBoundary component migration to Next.js chore: deleted test files
