--- rs-react-app/src/components/Error/ErrorBoundary.test.tsx
@@ -1,70 +1,0 @@
-import { render, screen } from '@testing-library/react';
-import userEvent from '@testing-library/user-event';
-import ErrorBoundary from './ErrorBoundary';
-import { vi } from 'vitest';
-
-const ErrorComponent = ({ shouldThrow }: { shouldThrow: boolean }) => {
-  if (shouldThrow) {
-    throw new Error('Test error');
-  }
-  return <div>No error</div>;
-};
-
-describe('ErrorBoundary', () => {
-  beforeEach(() => {
-    vi.spyOn(console, 'error').mockImplementation(() => {});
-  });
-
-  afterEach(() => {
-    vi.restoreAllMocks();
-  });
-
-  it('catches error and displays fallback UI when child throws', () => {
-    render(
-      <ErrorBoundary>
-        <ErrorComponent shouldThrow={true} />
-      </ErrorBoundary>
-    );
-
-    expect(screen.getByText('Something went wrong.')).toBeInTheDocument();
-    expect(
-      screen.getByText('Please try reloading the page.')
-    ).toBeInTheDocument();
-    expect(
-      screen.getByRole('button', { name: /reload page/i })
-    ).toBeInTheDocument();
-  });
-
-  it('renders children normally when no error occurs', () => {
-    render(
-      <ErrorBoundary>
-        <ErrorComponent shouldThrow={false} />
-      </ErrorBoundary>
-    );
-
-    expect(screen.getByText('No error')).toBeInTheDocument();
-    expect(screen.queryByText('Something went wrong.')).not.toBeInTheDocument();
-  });
-
-  it('reloads page when reload button is clicked', async () => {
-    const mockReload = vi.fn();
-    Object.defineProperty(window, 'location', {
-      value: { reload: mockReload },
-      writable: true,
-    });
-
-    render(
-      <ErrorBoundary>
-        <ErrorComponent shouldThrow={true} />
-      </ErrorBoundary>
-    );
-
-    const user = userEvent.setup();
-    const reloadButton = screen.getByRole('button', { name: /reload page/i });
-    await user.click(reloadButton);
-
-    expect(mockReload).toHaveBeenCalledTimes(1);
-
-    mockReload.mockRestore();
-  });
-});
