+++ rs-react-app/src/styles/error.module.css
